restructure and added part of ogres key quest
